+++ run.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-compdd validate_run_vina --config sample_configs/vina_validation.yaml 
-compdd validate_run_dock6 --config sample_configs/dock6_validation.yaml