+++ src/compdd/validation_coreset/validation_config.py
@@ -1,5 +1,5 @@
-from compdd.configs.root_config import RootConfig, _setup_dirs
-from compdd.configs.config_helpers import validate_and_normalize_receptors
+from compdd.docking_configs.root_config import RootConfig, _setup_dirs
+from compdd.docking_configs.config_helpers import validate_and_normalize_receptors
 from compdd.utils.extract_files import extract_files
 from pathlib import Path
 
@@ -11,9 +11,13 @@ def load_validation_config(path):
     cfg = RootConfig.model_validate(data)
 
     data = Path(cfg.validation.data)
-    pdbs = extract_files(data, "_protein.pdb", recursive=True)
-    reference = extract_files(data, "_pocket.pdb", recursive=True)
-    sdfs = extract_files(data, "_ligand.sdf", recursive=True)
+    protein_suffix = cfg.validation.protein_suffix
+    pocket_suffix = cfg.validation.pocket_suffix
+    ligand_suffix = cfg.validation.ligand_suffix
+
+    pdbs = extract_files(data, protein_suffix, recursive=True)
+    reference = extract_files(data, pocket_suffix, recursive=True)
+    sdfs = extract_files(data, ligand_suffix, recursive=True)
 
     cfg.receptors.source = "pdb"
     cfg.receptors.pdbs = pdbs